Misc cleanup
Improve the handling of pending jumps with a specific enum that records the to-be-resolved jump.

Reduce binops to dual registers again, they just don't use the dst register enough for it to make sense.

Make the foreach loop use less registers by moving the initializers for the loop variables directly into the target immediate register rather than allocating an intermediary