Cleanup handling of ident -> reg mapping
Store the previous mapping in the list of declared locals, only keeping the first replaced mapping (since all others must be shadows in the same scope and don't need to be restored).
Don't store globals in the locals list, just do a lookup in the global list if a visible local entry isn't found.